backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
py-vollib==1.0.1
vnstock==0.2.8
requests==2.31.0
python-dotenv==1.0.0 